+++ rollup.config.js
@@ -4,25 +4,25 @@ import commonjs from '@rollup/plugin-commonjs';
 import resolve from '@rollup/plugin-node-resolve';
 import {terser} from 'rollup-plugin-terser';
 import typescript from '@rollup/plugin-typescript';
-// import vue from 'rollup-plugin-vue';
+import vue from 'rollup-plugin-vue';
 
 export default {
-  input: 'src/App.vue',
+  input: 'src/main.ts',
   output: {
     dir: 'dist',
     format: 'esm',
-    manualChunks: {
-      react: ['react'],
-      reactDOM: ['react-dom'],
-    },
     plugins: [terser()],
   },
+  external: ['vue'],
   plugins: [
     // brotli(),
     // replace({
     //   'process.env.NODE_ENV': JSON.stringify('production')
     // }),
-    typescript(),
+    typescript({
+      tsconfig: false,
+    }),
+    vue(),
     resolve(),
     commonjs({
       include: 'node_modules/**'